データのフェッチング箇所を変更
